+++ test/clevertap_signedcall_flutter_method_channel_test.dart
@@ -4,8 +4,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  MethodChannelClevertapSignedCallFlutter platform =
-      MethodChannelClevertapSignedCallFlutter();
+  MethodChannelCleverTapSignedCallFlutter platform =
+      MethodChannelCleverTapSignedCallFlutter();
   const MethodChannel channel = MethodChannel(channelName);
 
   TestWidgetsFlutterBinding.ensureInitialized();