+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## CHANGE LOG
 
-* TODO: Describe initial release.
+### Version 0.0.1 (, 2022)
+* Initial Release.
+* Supports Signed Call Android SDK v0.0.1
+* Supports Signed Call Android iOS v0.0.1